Remove drive-by changes from project session cutover
