+++ src/components/layout/areas/MemoryPane.tsx
@@ -1,19 +1,22 @@
 import ReactJsonView from '@microlink/react-json-view'
 import toast from 'react-hot-toast'
 
+import type { ExtrudeSurface } from '@rust/kcl-lib/bindings/ExtrudeSurface'
+import type { Path } from '@rust/kcl-lib/bindings/Path'
+
 import { ActionButton } from '@src/components/ActionButton'
 import Loading from '@src/components/Loading'
 import Tooltip from '@src/components/Tooltip'
 import { LayoutPanel, LayoutPanelHeader } from '@src/components/layout/Panel'
 import { useModelingContext } from '@src/hooks/useModelingContext'
 import { useResolvedTheme } from '@src/hooks/useResolvedTheme'
+import type { VariableMap } from '@src/lang/wasm'
+import { humanDisplayNumber, sketchFromKclValueOptional } from '@src/lang/wasm'
 import { useSingletons } from '@src/lib/boot'
 import type { AreaTypeComponentProps } from '@src/lib/layout'
-import { trap } from '@src/lib/trap'
+import { Reason, trap } from '@src/lib/trap'
+import type { ModuleType } from '@src/lib/wasm_lib_wrapper'
 import { Suspense, use } from 'react'
-import { processMemory } from '@src/components/layout/areas/MemoryPane.utils'
-
-export { processMemory }
 
 export const MemoryPaneMenu = () => {
   const { kclManager } = useSingletons()
@@ -104,4 +107,43 @@ export const MemoryPaneContents = () => {
       )}
     </div>
   )
+}
+
+export const processMemory = (
+  variables: VariableMap,
+  wasmInstance: ModuleType
+) => {
+  const processedMemory: Record<
+    string,
+    string | number | boolean | object | undefined
+  > = {}
+  for (const [key, val] of Object.entries(variables)) {
+    if (val === undefined) continue
+    const sk = sketchFromKclValueOptional(val, key)
+    if (val.type === 'Solid') {
+      processedMemory[key] = val.value.value.map(
+        ({ ...rest }: ExtrudeSurface) => {
+          return rest
+        }
+      )
+    } else if (!(sk instanceof Reason)) {
+      processedMemory[key] = sk.paths.map(({ __geoMeta, ...rest }: Path) => {
+        return rest
+      })
+    } else if (val.type === 'Function') {
+      processedMemory[key] = '__function__'
+    } else if (val.type === 'Number') {
+      processedMemory[key] = humanDisplayNumber(val.value, val.ty, wasmInstance)
+    } else if (val.type === 'SketchVar') {
+      const sketchVar = val.value
+      processedMemory[key] =
+        `var ${humanDisplayNumber(sketchVar.initialValue, sketchVar.ty, wasmInstance)}`
+    } else if (val.type === 'Enum') {
+      // Enums are shown by nominal identity, the same way they are written.
+      processedMemory[key] = `${val.enum_name}::${val.variant}`
+    } else {
+      processedMemory[key] = val.value
+    }
+  }
+  return processedMemory
 }
